# jouko-api-spec/jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Jouko API spec",
    "version" : "0.0.4",
    "title" : "Jouko API"
  },
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
    "/admin/interruptiongroups/" : {
      "get" : {
        "tags" : [ "Interruption Groups" ],
        "summary" : "List interruption groups",
        "operationId" : "listInterruptionGroups",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "The offset of the first result",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "The maximum number of results",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/InterruptionGroup"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Interruption Groups" ],
        "summary" : "Create interruption group",
        "operationId" : "createInterruptionGroup",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "The cancellation status",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/InterruptionGroup"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/InterruptionGroup"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/admin/interruptiongroups/{groupId}" : {
      "get" : {
        "tags" : [ "Interruptions" ],
        "summary" : "Retreive interruption group",
        "operationId" : "retrieveInterruptionGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "description" : "The id of the interruption group",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/InterruptionGroup"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Interruptions" ],
        "summary" : "Update interruption group",
        "operationId" : "updateInterruptionGroup",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "description" : "The id of the interruption group",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "The cancellation status",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/InterruptionGroup"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/InterruptionGroup"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Interruption Groups" ],
        "summary" : "Delete interruption",
        "description" : "Deletes an interruption",
        "operationId" : "deleteInterruption",
        "parameters" : [ {
          "name" : "groupId",
          "in" : "path",
          "description" : "The id of the interruption being deleted",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/admin/controllerDevices" : {
      "get" : {
        "tags" : [ "ControllerDevices" ],
        "summary" : "List all controller devices",
        "operationId" : "listAllControllerDevices",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "The offset of the first result",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "The maximum number of results",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ControllerDevice"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "ControllerDevices" ],
        "summary" : "Create controller device",
        "operationId" : "createControllerDevice",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "The controller device to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ControllerDevice"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ControllerDevice"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/admin/controllerDevices/{controllerDeviceId}" : {
      "get" : {
        "tags" : [ "ControllerDevices" ],
        "summary" : "Retreive controller device",
        "operationId" : "retrieveControllerDevice",
        "parameters" : [ {
          "name" : "controllerDeviceId",
          "in" : "path",
          "description" : "The id of the controller device",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ControllerDevice"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "ControllerDevices" ],
        "summary" : "Update controller device",
        "operationId" : "updateControllerDevice",
        "parameters" : [ {
          "name" : "controllerDeviceId",
          "in" : "path",
          "description" : "The id of the controller device",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "newControllerDevice",
          "description" : "The new data for the controller device",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ControllerDevice"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/ControllerDevice"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "ControllerDevices" ],
        "summary" : "Delete controller",
        "description" : "Deletes an controller",
        "operationId" : "deleteControllerDevice",
        "parameters" : [ {
          "name" : "controllerDeviceId",
          "in" : "path",
          "description" : "The id of the controller",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/admin/devices" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "List all devices",
        "operationId" : "listAllDevices",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "The offset of the first result",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "The maximum number of results",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Device"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Devices" ],
        "summary" : "Create device",
        "operationId" : "createDevice",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "The device to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Device"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Device"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/admin/devices/{deviceId}" : {
      "get" : {
        "tags" : [ "Devices" ],
        "summary" : "Retreive device",
        "operationId" : "retrieveDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "The id of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Device"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Devices" ],
        "summary" : "Update device",
        "operationId" : "updateDevice",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "The id of the device",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "newDevice",
          "description" : "The new data for the device",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Device"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/Device"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/admin/users" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "List all users",
        "operationId" : "listAllUsers",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "The offset of the first result",
          "required" : true,
          "type" : "integer"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "The maximum number of results",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Users" ],
        "summary" : "Create user",
        "operationId" : "createUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "The user to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        }, {
          "name" : "token",
          "in" : "query",
          "description" : "Keycloak token",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "Unauthorized" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
    "InternalServerError" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
...
    "InterruptionGroup" : {
      "type" : "object",
      "required" : [ "endTime", "id", "startTime" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Interruption id"
        },
        "startTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Start time of the interruption, inclusive"
        },
        "endTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "End time of the interruption, exclusive"
        },
        "overbookingFactor" : {
          "type" : "number",
          "format" : "double",
          "description" : "Overbooking factor"
        },
        "powerSavingGoalInWatts" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "The amount of power this interruption is supposed to save, in watts"
        }
      }
    },
...
    "ControllerDevice" : {
      "type" : "object",
      "required" : [ "eui", "key" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "Controller device id"
        },
        "eui" : {
          "type" : "string",
          "description" : "The EUI of the controller device"
        },
        "key" : {
          "type" : "string",
          "description" : "The key of the controller device"
        },
        "communicationChannel" : {
          "type" : "string",
          "description" : "The communication channel of the controller device"
        }
      }
    },
...
  }
}